requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==1.31.5
google-auth==1.35.0
google-cloud==0.34.0
google-cloud-bigquery==2.4.0
google-cloud-core==1.7.2
google-cloud-dataproc==4.0.1
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==1.3.3
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
numpy==1.22.3
packaging==21.3
pandas==1.4.2
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
